use the newest Twilio lib beta9 + remove pods from git
- add instruction on how to add podfile to your project
- rename the ios module from TwilioVoice to RNTwilioVoide because of name clash with Twilio
- use weak Framework reference for TwilioVoice.framework
