+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.hoxfon.react.TwilioVoice">
+    package="com.hoxfon.react.RNTwilioVoice">
     <uses-permission android:name="android.permission.DISABLE_KEYGUARD"/>
 </manifest>
