+++ ios/RNTwilioVoice/RNTwilioVoice.h
@@ -1,0 +1,11 @@
+//
+//  TwilioVoice.h
+//  TwilioVoice
+//
+
+#import <React/RCTBridgeModule.h>
+#import <React/RCTEventEmitter.h>
+
+@interface RNTwilioVoice : RCTEventEmitter <RCTBridgeModule>
+
+@end
