--- ios/TwilioVoice/TwilioVoice.h
@@ -1,14 +1,0 @@
-//
-//  TwilioVoice.h
-//  TwilioVoice
-//
-//  Created by Erva on 3/20/17.
-//  Copyright © 2017 Erva. All rights reserved.
-//
-
-#import <React/RCTBridgeModule.h>
-#import <React/RCTEventEmitter.h>
-
-@interface TwilioVoice : RCTEventEmitter <RCTBridgeModule>
-
-@end
